Reverted certain parts
* Added structure

* Finished EDA

* Added requirement packages file

* Splitted dataset

* Finished model training section

* Finished Model evaluation

* Updated gitignore

* Improved model by fixing the transformer

* Created two types of model training

* Finalized for the API

* Updated for streamlit

* Added files for the UI

* Added file for the form

* Added

* Updated get prediction

* Updated with the past predictions

* Reorganized folder

* Updated for final touch before the follow up

* Removed .idea folder

* Renamed folder

* Updated from last session

* Updated for airflow

* Updated for pydantic

* Added splitted files script

* Added splitted files for jobs

* Added dag file for prediction

* Updated dag file with scheduled prediction

* Updated for dag

* Added for data validation

* Updated DAG to split into folders

* Updated with libraries

* Updated prediction job

* Removed wrong files

* Added csv files for processing

* Updated with the correct expectations

* Added airflow config files

* Updated with correct names

* Updated validation logics

* Updated for merge

* Updated scheduled predictions

* Updated requirements file

* Updated for the DAG

* Added docker file for airflow

* Updated for final submission

* Removed unnecessary files

* For merge

* For merge

* For merge

* For merge 2

* For merge

* Test merge

* Test merge

* Updated codes quality

* Removed unnecessary files

* Added to make api being regconized

* Final version of custom image

* Added image for readme

* Updated README

* Added files for the generation of dummy files

* Updated README with csv generation

* Updated quality

* Updated quality

* Updated quality

* Reverted naming

* Committed for merge

* Not included

* Updated README

---------

+++ api/models.py
@@ -44,7 +44,7 @@ class Customer(Base):
     CustomerId = Column(Integer, autoincrement=True, primary_key=True)
     CreditScore = Column(Integer)
     Gender = Column(String)
-    Age = Column(Integer)
+    Age = Column(Float)
     Tenure = Column(Integer)
     Balance = Column(Float)
     NumOfProducts = Column(Integer)
